requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-environ==0.4.4
django==2.0.3
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2==2.5.4
pytz==2018.3              # via django
six==1.11.0               # via django-environ
whitenoise==1.0.6
